fix: renamed kebab-cased ts files to camelCase.
