+++ src/dev/versionPlugin.ts
