+++ src/lib/composables/index.ts
@@ -1,3 +1,3 @@
-export * from "./use-align";
-export * from "./use-router-link";
-export * from "./use-size";
+export * from "./useAlign";
+export * from "./useRouterLink";
+export * from "./useSize";
